Opprettet ny service AutomatiskBeslutningService for å unngå sirkulær avhengighet
