+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/steg/BeslutteVedtakTest.kt
@@ -13,6 +13,7 @@ import no.nav.familie.ba.sak.common.lagVedtak
 import no.nav.familie.ba.sak.config.FeatureToggleConfig
 import no.nav.familie.ba.sak.config.FeatureToggleService
 import no.nav.familie.ba.sak.config.TaskRepositoryWrapper
+import no.nav.familie.ba.sak.kjerne.behandling.AutomatiskBeslutningService
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingStatus
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
@@ -52,6 +53,7 @@ class BeslutteVedtakTest {
     private lateinit var featureToggleService: FeatureToggleService
     private lateinit var tilkjentYtelseValideringService: TilkjentYtelseValideringService
     private lateinit var simuleringService: SimuleringService
+    private lateinit var automatiskBeslutningService: AutomatiskBeslutningService
 
     private val saksbehandlerContext = mockk<SaksbehandlerContext>()
 
@@ -69,6 +71,7 @@ class BeslutteVedtakTest {
         featureToggleService = mockk()
         tilkjentYtelseValideringService = mockk()
         simuleringService = mockk()
+        automatiskBeslutningService = mockk()
 
         val loggService = mockk<LoggService>()
 
@@ -104,7 +107,7 @@ class BeslutteVedtakTest {
             featureToggleService,
             tilkjentYtelseValideringService,
             saksbehandlerContext,
-            simuleringService,
+            automatiskBeslutningService,
         )
     }
