+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/steg/StegServiceTest.kt
@@ -41,8 +41,8 @@ internal class StegServiceTest {
         tilgangService = mockk(relaxed = true),
         infotrygdFeedService = mockk(),
         satsendringService = satsendringService,
-        simuleringService = simuleringService,
         personopplysningerService = mockk(),
+        automatiskBeslutningService = mockk(),
     )
 
     @BeforeEach
